requirements.txt
fastapi
uvicorn
openai
pytesseract
pillow
python-multipart
requests
tqdm
faiss-cpu
python-dotenv
numpy
